Refactor - prevent jumps between sections of points
As a limitation of GoogleMaps api for Directions, it's only possible to
add 8 waypoints between origin and destination.
To accomplish that it was necessary slice the dataset accordingly but at
first it was made without connecting the destination of the last slice
with the origin of the next one... letting have jumps between one slice
to other...

This commit prevent it using as origin always the last destination of
the previous slice.